+++ src/main/java/com/roladio/banking/dto/TransferRequest.java
@@ -1,3 +1,5 @@
 package com.roladio.banking.dto;
 
-public record TransferRequest(Long fromId, Long toId, double amount) {}
+import java.math.BigDecimal;
+
+public record TransferRequest(Long fromId, Long toId, BigDecimal amount) {}
